Offer to open workspace in root even when telemetry has been disabled (fix #108669)
